Update Tiny_Monster_Trainer.py
Changed combat math so that there will be less misses at the start of the game. ( And hopefully it scales up okay? :D My fear was that after a certain point you wouldn't be able to hit a high agility monster, at all. Now I hope I've fixed it so that I don't have to deal with that nightmare. Now I just hope that I didn't go too far the other way? But that's better than not being able hit something at all)
Added "Glance!" for when an attack would be a miss, but the attacker still finds a way to connect the blow. A glance will do half as much damage rounded up instead of full damage.
made the buttonInput function shorter